Extracting simulation visualisation from LandingComponent
