+++ demo/app.module.ts
@@ -1,6 +1,8 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { LandingComponent }  from '../scripts/components/app.landing.component';
+import { TaktTimeComponent }  from '../scripts/components/app.tt.component';
+import { SimulationComponent }  from '../scripts/components/app.simulation.component';
 
 import { SimulationService } from '../scripts/services/simulation.service';
 import { WorkItemsService } from '../scripts/services/work-items.service';
@@ -12,12 +14,14 @@ import { MockWorkItemTrackingClient } from './mocks/work-item-tracking.client';
 @NgModule({
   imports: [BrowserModule],
   declarations: [
-                LandingComponent
-                ],
-  bootstrap: [
-                LandingComponent
-             ],
-  providers: [
+        LandingComponent,
+        TaktTimeComponent,
+        SimulationComponent
+    ],
+    bootstrap: [
+        LandingComponent
+    ],
+    providers: [
         SimulationService,
         WorkItemsService,
         {
