+++ scripts/Components/app.landing.component.html
@@ -1,10 +1,7 @@
 <div class="hub-content">
     <div class="splitter horizontal hub-splitter stateful toggle-button-enabled toggle-button-hotkey-enabled">
-    <div class="hub-title">Hello Monte Carlo</div>
-    <div class="right-hub-content">        
-        <div id="forecast_matrix"></div>
-        <div>Completed work item count: {{CompletedWorkItemsCount}}</div>
-        <div>In progress work item count: {{InProgressWorkItemsCount}}</div>
-    </div>
-    </div>
+        <div class="hub-title">Hello Monte Carlo</div>
+        <simulation></simulation>
+        <takt-time></takt-time>
+    </div>    
 </div>
