+++ scripts/components/app.simulation.component.html
@@ -1,0 +1,5 @@
+<div class="right-hub-content">        
+    <div id="forecast_matrix"></div>
+    <div>Completed work item count: {{CompletedWorkItemsCount}}</div>
+    <div>In progress work item count: {{InProgressWorkItemsCount}}</div>
+</div>
